pyproject.toml
[project]
name = "MPM"
version = "1.0"
dependencies = ["cachetools==5.4.0",
"certifi==2024.7.4",
"cffi==1.16.0",
"charset-normalizer==3.3.2",
"cryptography==43.0.0",
"google-api-core==2.19.1",
"google-api-python-client==2.139.0",
"google-auth==2.32.0",
"google-auth-httplib2==0.2.0",
"google-auth-oauthlib==1.2.1",
"googleapis-common-protos==1.63.2",
"httplib2==0.22.0",
"idna==3.7",
"oauthlib==3.2.2",
"pillow==10.4.0",
"proto-plus==1.24.0",
"protobuf==5.27.3",
"pyasn1==0.6.0",
"pyasn1_modules==0.4.0",
"pycparser==2.22",
"pyparsing==3.1.2",
"pyperclip==1.9.0",
"requests==2.32.3",
"requests-oauthlib==2.0.0",
"rsa==4.9",
"uritemplate==4.1.1",
"urllib3==2.2.2"]